requirements.txt
Flask==3.0.3
openai
requests
azure-search-documents
flask